Small code improvements and bug fixes
Removed JSONAPI Utils from BM Controller, as well as the show action.
Added a `with_private` scope in the BME to unscope the default scope.
Changed the `bmes` relationship in BM to use the with `with_private`
scope.
Added the relationships to BMEs and BMs in `BusinessModelBmeResource`.
Added the primary key and verify key to BM Resource.